requirements.txt
rich>=13.7.0
pyyaml>=6.0.1
click>=8.1.7
jsonschema>=4.21.0
